Kludge: Copied lab from 131 folder to 125
This appears to fix a weird bug.  When a font is first used in a remote
file (using subinclude, from the import package), the font is not found.
So I've made a point of loading this one lab locally, so that all other
labs that use the same fonts work fine remotely.