Added directories and dummy files
Mac Tex installation can't create directories when it needs to to house the .aux files for included labs. This causes errors when using \includelab for labs that live in other directories. To fix this, I created all of the needed directories. To make sure GitHub maintains this directory structure, I put a dummy file in each.